+++ changelog.md
@@ -1,6 +1,7 @@
-## version 0.8.1 -> 0.8.6
+## version 0.8.1 -> 0.8.9
  - Added new automatic Docker mapping feature (for people not using (sub)domains)
  - Added guardrails to prevent Let's Encrypt from failing to initialize when adding wrong domains
+ - Add search bar on the marketplace
  - App store image size issue
  - Fixed wrong x-forwarded-proto header 
  - Add installer option for hostname prefix/suffix